make sure telemetry proc exits; address comments
The background process that uploads telemetry was hanging when the
telemetry keys weren't set. This was due to a goroutine returning
before calling `WaitGroup.Done`. Add a defer for the `WaitGroup.Done`
call and add a new goroutine that calls `os.Exit` if the process
doesn't exit after 5 seconds.
