+++ internal/boxcli/log.go
@@ -7,7 +7,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"go.jetpack.io/devbox/internal/boxcli/usererr"
-	"go.jetpack.io/devbox/internal/envir"
 	"go.jetpack.io/devbox/internal/telemetry"
 )
 
@@ -34,14 +33,14 @@ func doLogCommand(cmd *cobra.Command, args []string) error {
 			return usererr.New("expected a start-time argument for logging the shell-ready event")
 		}
 		telemetry.Event(telemetry.EventShellReady, telemetry.Metadata{
-			CommandStart: envir.ParseShellStart(args[1]),
+			CommandStart: telemetry.ParseShellStart(args[1]),
 		})
 	case "shell-interactive":
 		if len(args) < 2 {
 			return usererr.New("expected a start-time argument for logging the shell-interactive event")
 		}
 		telemetry.Event(telemetry.EventShellInteractive, telemetry.Metadata{
-			CommandStart: envir.ParseShellStart(args[1]),
+			CommandStart: telemetry.ParseShellStart(args[1]),
 		})
 	}
 	return usererr.New("unrecognized event-name %s for command: %s", args[0], cmd.CommandPath())
