+++ internal/cloud/cloud.go
@@ -28,6 +28,7 @@ import (
 	"go.jetpack.io/devbox/internal/debug"
 	"go.jetpack.io/devbox/internal/envir"
 	"go.jetpack.io/devbox/internal/services"
+	"go.jetpack.io/devbox/internal/telemetry"
 	"go.jetpack.io/devbox/internal/ux/stepper"
 )
 
@@ -444,7 +445,7 @@ func shell(username, hostname, projectDir string, shellStartTime time.Time) erro
 	cmd := &openssh.Cmd{
 		DestinationAddr: hostname,
 		PathInVM:        absoluteProjectPathInVM(username, projectPath),
-		ShellStartTime:  envir.FormatShellStart(shellStartTime),
+		ShellStartTime:  telemetry.FormatShellStart(shellStartTime),
 		Username:        username,
 	}
 	sessionErrors := newSSHSessionErrors()
