+++ internal/envir/util.go
@@ -6,7 +6,6 @@ package envir
 import (
 	"os"
 	"strconv"
-	"time"
 )
 
 func IsDevboxCloud() bool {
@@ -37,28 +36,6 @@ func IsInBrowser() bool { // TODO: a better name
 func IsCI() bool {
 	ci, err := strconv.ParseBool(os.Getenv("CI"))
 	return ci && err == nil
-}
-
-func ShellStart() time.Time {
-	return ParseShellStart(os.Getenv(DevboxShellStartTime))
-}
-
-func FormatShellStart(t time.Time) string {
-	if t.IsZero() {
-		return ""
-	}
-	return strconv.FormatInt(t.Unix(), 10)
-}
-
-func ParseShellStart(s string) time.Time {
-	if s == "" {
-		return time.Time{}
-	}
-	unix, err := strconv.ParseInt(s, 10, 64)
-	if err != nil {
-		return time.Time{}
-	}
-	return time.Unix(unix, 0)
 }
 
 // GetValueOrDefault gets the value of an environment variable.
